# openapi.yaml
openapi: 3.1.0
info:
  title: Quits API
  version: 1.0.0
  summary: Backend contract expected by the Quits frontend.
  description: |
    The HTTP backend that `frontend/src/services/httpBackend.js` is written
    against. It is derived from that client and from the in-browser reference
    implementation in `frontend/src/services/mockBackend.js`, which enforces
    every rule below.

    `_docs/SPEC.md` is authoritative. Section references below point into it.

    ## Authentication

    There is none, in the usual sense, and this is deliberate (SPEC §2).

    - There are no accounts, no login, no sessions, no passwords and no roles.
    - **Possession of the trip token is the only credential.** It is a path
      segment, not a header or a cookie. The client sends no `Authorization`
      header, no cookie and no API key on any request.
    - Anyone holding a trip's token has full read and write access to that
      trip, including editing and deleting expenses they did not enter.
    - The trip creator has no elevated rights after creation.

    Consequently there are exactly two access tiers:

    | Endpoint | Credential |
    |---|---|
    | `POST /api/trips` | None. Fully public — this is how a token is obtained. |
    | Everything under `/api/trips/{token}` | The `token` path parameter. |

    This is **not** expressible as an OpenAPI `securityScheme`: `apiKey`
    supports `header`, `query` and `cookie`, but not `path`. Declaring one
...
    `/api/trips/{token}`.

...

...

...

...

...

...

...

...
servers:
...
paths:
  /api/trips:
    post:
      tags: [Trips]
      operationId: createTrip
      security: []
      summary: Create a trip
      description: |
        SPEC §7.1. The only public endpoint: it requires no credential, because
        it is what issues one.

        The creator becomes participant #1. They do not pre-list anyone else —
        everyone else joins through the returned link (SPEC §3).

        The response carries the trip's secret `token`. It is returned exactly
        once here and on subsequent reads by anyone who already holds it; there
        is no recovery path if the caller loses it (SPEC §2).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTripRequest'
            example:
              name: Lisbon, March
              currencyLabel: EUR
              creatorName: Anna
      responses:
        '201':
          description: |
            Trip created. The client accepts any 2xx.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTripResponse'
        '422':
          description: |
            `trip_name_empty` — `name` is empty after trimming.
            `name_empty` — `creatorName` is empty after trimming.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/trips/{token}:
    parameters:
      - $ref: '#/components/parameters/Token'
    get:
      tags: [Trips]
      operationId: getTrip
      security: []
      summary: Read a trip in full
      description: |
        SPEC §7.2. **Authorization: the `token` path parameter.**

        Returns everything the app renders in one document: the trip,
        participants, expenses with their resolved splits, and the computed
        balances, transfers and total.

        Balances and transfers are computed server-side on every read and are
        never stored (SPEC §6). The frontend does not recompute them.

        No identity is required. A visitor who has not joined sees the trip
        read-only and is asked for a name only when they try to change
        something (SPEC §11.5).
      responses:
        '200':
          description: The trip and everything in it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripView'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/trips/{token}/participants:
    parameters:
      - $ref: '#/components/parameters/Token'
    post:
      tags: [Participants]
      operationId: joinTrip
      security: []
      summary: Join a trip under a new name
      description: |
        SPEC §7.5. **Authorization: the `token` path parameter.**

        Creates a new participant. Names are unique per trip, compared
        case-insensitively after trimming.

        A taken name is rejected outright. Do not fuzzy-match and do not ask
        "is this you?" — the message should tell the caller to add a
        distinguishing suffix. Someone who *is* the existing participant, on a
        second device, uses `POST /participants/claim` instead (SPEC §3).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameRequest'
            example:
              name: Max
      responses:
        '201':
          description: Participant created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
        '409':
          description: |
            `name_taken` — another participant in this trip already uses that
            name. `details.field` is `name`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: '`name_empty` — the name is empty after trimming.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/trips/{token}/participants/claim:
    parameters:
      - $ref: '#/components/parameters/Token'
    post:
      tags: [Participants]
      operationId: claimParticipant
      security: []
      summary: Bind to an existing participant from another device
      description: |
        SPEC §7.6 and §3. **Authorization: the `token` path parameter.**

        The "this is me on another device" path. Looks up an existing
        participant by name — case-insensitive, trimmed — and returns it
        **without creating anything**. This is also how someone whose browser
        storage was cleared recovers their identity.

        This endpoint is read-only despite being a POST: it must never create a
        participant, and it does not update `lastActivityAt`.

        Note for routing: register this literal path before any pattern that
        could capture `claim` as a participant id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameRequest'
            example:
              name: Max
      responses:
        '200':
          description: The existing participant with that name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
        '404':
          description: |
            `not_found` — either the trip token is unknown, or nobody in the
            trip uses that name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: '`name_empty` — the name is empty after trimming.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/trips/{token}/participants/{participantId}:
    parameters:
      - $ref: '#/components/parameters/Token'
      - $ref: '#/components/parameters/ParticipantId'
    patch:
      tags: [Participants]
      operationId: renameParticipant
      security: []
      summary: Rename a participant
      description: |
        SPEC §7.7. **Authorization: the `token` path parameter.**

        Anyone with the link may rename anyone. Under this join model renaming
        is routine, not an edge case.

        The same uniqueness check as joining applies, except against the
        participant's own current name — changing only the casing of your own
        name is allowed.

        Renaming never detaches expenses: they reference the participant id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameRequest'
            example:
              name: Anna K
      responses:
        '200':
          description: The renamed participant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
        '409':
          description: '`name_taken` — another participant already uses that name.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: '`name_empty` — the name is empty after trimming.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: '`not_found` — unknown trip token, or no such participant in this trip.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Participants]
      operationId: deleteParticipant
      security: []
      summary: Remove a participant
      description: |
        SPEC §7.8. **Authorization: the `token` path parameter.**

        Permitted **only** when the participant has zero expenses as payer and
        zero split rows. This is the mechanism for removing someone who joined
        out of curiosity and never took part; it is not a way to rewrite
        history, because removing someone who is on an expense would silently
        change what everyone else owes.

        When refused, `details` carries `paidCount` and `splitCount` so the
        client can explain why. The frontend also disables the control using
        the same counts derived from the trip view, but the server is the gate.
      responses:
        '204':
          description: Removed. No body.
        '409':
          description: |
            `participant_in_use` — the participant is on at least one expense.
            `details.paidCount` and `details.splitCount` say how.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: '`not_found` — unknown trip token, or no such participant in this trip.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/trips/{token}/expenses:
    parameters:
      - $ref: '#/components/parameters/Token'
    post:
      tags: [Expenses]
      operationId: createExpense
      security: []
      summary: Add an expense
      description: |
        SPEC §7.10. **Authorization: the `token` path parameter.**

        The split is resolved to integer cents at save time and those cents are
        what get stored. `splitType` records only how the rows were produced;
        nothing downstream reads it (SPEC §4).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseInput'
            examples:
              equal:
                summary: Split equally between three people
                value:
                  description: Dinner at Ramiro
                  amountCents: 10000
                  paidBy: p_3ijg02LqmzmW
                  date: '2026-09-20'
                  category: food
                  splitType: equal
                  participantIds: [p_3ijg02LqmzmW, p_iTk6k0Zl_U0z, p_9aQ2vB1xKpLm]
                  exactCents: {}
                  basisPoints: {}
                  createdBy: p_3ijg02LqmzmW
              payerOutsideSplit:
                summary: Anna buys the ferry tickets and does not go
                value:
                  description: Ferry tickets
                  amountCents: 1700
                  paidBy: p_3ijg02LqmzmW
                  date: '2026-09-20'
                  category: transport
                  splitType: equal
                  participantIds: [p_iTk6k0Zl_U0z, p_9aQ2vB1xKpLm]
                  exactCents: {}
                  basisPoints: {}
                  createdBy: p_3ijg02LqmzmW
              percent:
                summary: Percent split, as integer basis points
                value:
                  description: Groceries
                  amountCents: 10000
                  paidBy: p_iTk6k0Zl_U0z
                  date: '2026-09-20'
                  category: null
                  splitType: percent
                  participantIds: [p_3ijg02LqmzmW, p_iTk6k0Zl_U0z, p_9aQ2vB1xKpLm]
                  exactCents: {}
                  basisPoints:
                    p_3ijg02LqmzmW: 3333
                    p_iTk6k0Zl_U0z: 3333
                    p_9aQ2vB1xKpLm: 3334
                  createdBy: p_3ijg02LqmzmW
      responses:
        '201':
          description: The stored expense, with its resolved splits.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '422':
          $ref: '#/components/responses/ExpenseValidationFailed'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/trips/{token}/expenses/{expenseId}:
    parameters:
      - $ref: '#/components/parameters/Token'
      - $ref: '#/components/parameters/ExpenseId'
    put:
      tags: [Expenses]
      operationId: updateExpense
      security: []
      summary: Replace an expense
      description: |
        SPEC §7.15. **Authorization: the `token` path parameter.**

        A full replacement, not a patch: every field must be sent. The splits
        are recomputed and replaced on every save, including when `splitType`
        changes. Participants dropped from `participantIds` lose their split
        rows entirely.

        Anyone with the link may edit any expense, including one they did not
        enter (SPEC §2).

        **`createdBy` is accepted but ignored.** The client sends the current
        user's id in the body because it reuses the create payload, but the
        original author must be preserved — the expense list shows
        "added by <name>" and editing is not authorship. Validate the field if
        you like; do not write it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseInput'
      responses:
        '200':
          description: The updated expense, with freshly resolved splits.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
        '422':
          $ref: '#/components/responses/ExpenseValidationFailed'
        '404':
          description: '`not_found` — unknown trip token, or no such expense in this trip.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Expenses]
      operationId: deleteExpense
      security: []
      summary: Delete an expense
      description: |
        SPEC §7.16. **Authorization: the `token` path parameter.**

        Anyone with the link may delete any expense, including ones they did
        not enter. This is deliberate (SPEC §2). There is no confirmation step
        in the API and no soft delete.

        Removing the expense must remove its split rows too.
      responses:
        '204':
          description: Deleted. No body.
        '404':
          description: '`not_found` — unknown trip token, or no such expense in this trip.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  parameters:
    Token:
      name: token
      in: path
      required: true
      description: |
        The trip's secret token, and the only credential in this API.
        Cryptographically random, at least 128 bits, URL-safe (SPEC §4). The
        client percent-encodes it before use.
      schema:
        type: string
        minLength: 22
      example: 9xQx0Zr4Kk2mVtJb7Hn3Lw5PcYeAfUdS
    ParticipantId:
      name: participantId
      in: path
      required: true
      description: Opaque participant identifier, as returned by the API.
      schema:
        type: string
      example: p_3ijg02LqmzmW
    ExpenseId:
      name: expenseId
      in: path
      required: true
      description: Opaque expense identifier, as returned by the API.
      schema:
        type: string
      example: e_7Kd1pQz0Rm4X
  responses:
    NotFound:
      description: |
        `not_found` — the trip token is unknown. Returned rather than a 401 or
        403, because possession of the token is the only credential and there
        is no other reason access could be denied.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    CreateTripRequest:
      type: object
      required: [name, currencyLabel, creatorName]
      properties:
        name:
          type: string
          description: Required, non-empty after trimming. Stored trimmed.
          example: Lisbon, March
        currencyLabel:
          type: string
          description: |
            Display string only. There is no conversion logic anywhere in the
            product (SPEC §1, §4). Blank or whitespace defaults to `EUR`. The
            frontend renders it as a prefix, e.g. `EUR 1,234.56`.
          default: EUR
          example: EUR
        creatorName:
          type: string
          description: |
            Required, non-empty after trimming. Becomes participant #1. The
            creator gains no elevated rights from this (SPEC §2).
          example: Anna
    CreateTripResponse:
      type: object
      required: [trip, participant]
      properties:
        trip:
          $ref: '#/components/schemas/Trip'
        participant:
          $ref: '#/components/schemas/Participant'
    NameRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: |
            Compared case-insensitively after trimming, and stored trimmed.
          example: Max
...
    Participant:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
          description: Unique within the trip, case-insensitive, trimmed.
        createdAt:
          type: string
          format: date-time
          description: |
            Defines the canonical participant ordering that the SPEC §5
            rounding tie-break depends on.
...
    Expense:
      type: object
      required:
        [id, description, amountCents, paidBy, date, category, splitType, createdBy, createdAt, splits]
      properties:
        id:
          type: string
        description:
          type: string
        amountCents:
          type: integer
          minimum: 1
        paidBy:
          type: string
          description: |
            One participant. **Not** constrained to be in the split — someone
            can pay for a meal they did not eat (SPEC §4).
        date:
          type: string
          format: date
          description: |
            `YYYY-MM-DD`. Any past date is accepted, and so is any future date.
            Defaults to today when omitted on write.
        category:
          oneOf:
            - $ref: '#/components/schemas/Category'
            - type: 'null'
          description: Optional. `null` when unset.
        splitType:
          $ref: '#/components/schemas/SplitType'
        createdBy:
          type: string
          description: |
            The participant who entered it, shown as "added by <name>". Set on
            create and never changed afterwards.
        createdAt:
          type: string
          format: date-time
        splits:
          type: array
          description: Resolved cents, in canonical participant order. Always sums to `amountCents`.
          items:
            $ref: '#/components/schemas/Split'
...
    TripView:
      type: object
      description: Everything the app needs to render a trip, in one response.
      required: [trip, participants, expenses, balances, transfers, totalCents]
      properties:
        trip:
          $ref: '#/components/schemas/Trip'
        participants:
          type: array
          description: Ascending `createdAt`. This order is load-bearing for rounding.
          items:
            $ref: '#/components/schemas/Participant'
        expenses:
          type: array
          description: Newest first, descending `createdAt` (SPEC §7.17).
          items:
            $ref: '#/components/schemas/Expense'
        balances:
          type: array
          description: One per participant, in the same order as `participants`.
          items:
            $ref: '#/components/schemas/Balance'
        transfers:
          type: array
          description: The minimised settlement list. Empty when everyone is square.
          items:
            $ref: '#/components/schemas/Transfer'
        totalCents:
          type: integer
          minimum: 0
          description: Sum of every expense in the trip.
    Error:
      type: object
      description: |
        The body of every non-2xx response. The client reads `code` to decide
        what to do and shows `message` to the user verbatim.
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: |
            Human-readable and safe to display as-is. The frontend does not
            rewrite it.
          example: '"Anna" is already taken in this trip. Add something to tell you apart, like "Anna B".'
        details:
          type: object
          description: |
            Optional, code-specific. Known keys:

            - `field` — which request field was at fault, so the client can
              highlight the input: `name`, `creatorName`, `description`,
              `amountCents`, `paidBy`, `createdBy`, `participantIds`,
              `exactCents`, `basisPoints`, `category`, `splitType`
            - `differenceCents` — with `exact_sum_mismatch`
            - `differenceBasisPoints` — with `percent_sum_mismatch`
            - `paidCount`, `splitCount` — with `participant_in_use`
            - `existingParticipantId` — with `name_taken`
          additionalProperties: true
    ErrorCode:
      type: string
      description: |
        Note that `network` is produced by the client when `fetch` itself
        fails. A server must never send it.
      enum:
        - not_found
        - trip_name_empty
        - name_empty
        - name_taken
        - description_empty
        - amount_not_positive
        - no_participants_in_split
        - exact_sum_mismatch
        - percent_sum_mismatch
        - participant_in_use
        - unknown_participant
        - invalid_category
        - invalid_split_type
        - network